ref(issues): Delete old external issues component
