--- static/app/components/group/externalIssuesList/streamlinedExternalIssueList.tsx
@@ -1,296 +1,0 @@
-import {Fragment} from 'react';
-import styled from '@emotion/styled';
-
-import {AlertLink} from '@sentry/scraps/alert';
-import {Button, LinkButton, type ButtonProps} from '@sentry/scraps/button';
-import {Tooltip} from '@sentry/scraps/tooltip';
-
-import {DropdownButton} from 'sentry/components/dropdownButton';
-import {DropdownMenu} from 'sentry/components/dropdownMenu';
-import ErrorBoundary from 'sentry/components/errorBoundary';
-import type {ExternalIssueAction} from 'sentry/components/group/externalIssuesList/hooks/types';
-import {useGroupExternalIssues} from 'sentry/components/group/externalIssuesList/hooks/useGroupExternalIssues';
-import {Placeholder} from 'sentry/components/placeholder';
-import {t} from 'sentry/locale';
-import type {Event} from 'sentry/types/event';
-import type {Group} from 'sentry/types/group';
-import type {Project} from 'sentry/types/project';
-import {trackAnalytics} from 'sentry/utils/analytics';
-import {useOrganization} from 'sentry/utils/useOrganization';
-
-function getActionLabelAndTextValue({
-  action,
-  integrationDisplayName,
-}: {
-  action: ExternalIssueAction;
-  integrationDisplayName: string;
-}): {label: string | React.JSX.Element; textValue: string} {
-  // If there's no subtext or subtext matches name, just show name
-  if (!action.nameSubText || action.nameSubText === action.name) {
-    return {
-      label: action.name,
-      textValue: action.name,
-    };
-  }
-
-  // If action name matches integration name, just show subtext
-  if (action.name === integrationDisplayName) {
-    return {
-      label: action.nameSubText,
-      textValue: `${action.name} ${action.nameSubText}`,
-    };
-  }
-
-  // Otherwise show both name and subtext
-  return {
-    label: (
-      <div>
-        <strong>{action.name}</strong>
-        <div>{action.nameSubText}</div>
-      </div>
-    ),
-    textValue: `${action.name} ${action.nameSubText}`,
-  };
-}
-
-interface ExternalIssueListProps {
-  event: Event;
-  group: Group;
-  project: Project;
-}
-
-export function StreamlinedExternalIssueList({
-  group,
-  event,
-  project,
-}: ExternalIssueListProps) {
-  const organization = useOrganization();
-  const {isLoading, integrations, linkedIssues} = useGroupExternalIssues({
-    group,
-    event,
-    project,
-  });
-
-  if (isLoading) {
-    return <Placeholder height="25px" testId="issue-tracking-loading" />;
-  }
-
-  const hasLinkedIssuesOrIntegrations = integrations.length || linkedIssues.length;
-  if (!hasLinkedIssuesOrIntegrations) {
-    return (
-      <AlertLink
-        variant="muted"
-        to={`/settings/${organization.slug}/integrations/?category=issue%20tracking`}
-      >
-        {t('Track this issue in Jira, GitHub, etc.')}
-      </AlertLink>
-    );
-  }
-
-  return (
-    <Fragment>
-      {linkedIssues.length > 0 && (
-        <IssueActionWrapper>
-          {linkedIssues.map(linkedIssue => (
-            <ErrorBoundary key={linkedIssue.key} mini>
-              <Tooltip
-                overlayStyle={{maxWidth: '400px'}}
-                position="bottom"
-                title={
-                  <LinkedIssueTooltipWrapper>
-                    <LinkedIssueName>{linkedIssue.title}</LinkedIssueName>
-                    <HorizontalSeparator />
-                    <UnlinkButton
-                      priority="link"
-                      size="zero"
-                      onClick={linkedIssue.onUnlink}
-                    >
-                      {t('Unlink issue')}
-                    </UnlinkButton>
-                  </LinkedIssueTooltipWrapper>
-                }
-                isHoverable
-              >
-                <LinkedIssue
-                  href={linkedIssue.url}
-                  external
-                  size="zero"
-                  icon={linkedIssue.displayIcon}
-                >
-                  <IssueActionName>{linkedIssue.displayName}</IssueActionName>
-                </LinkedIssue>
-              </Tooltip>
-            </ErrorBoundary>
-          ))}
-        </IssueActionWrapper>
-      )}
-      {integrations.length > 0 && (
-        <IssueActionWrapper>
-          {integrations.map(integration => {
-            const sharedButtonProps: ButtonProps = {
-              size: 'zero',
-              icon: integration.displayIcon,
-              priority: 'transparent',
-              children: <IssueActionName>{integration.displayName}</IssueActionName>,
-            };
-
-            if (integration.actions.length === 1) {
-              const action = integration.actions[0]!;
-              return (
-                <ErrorBoundary key={integration.key} mini>
-                  {action.href ? (
-                    // Exclusively used for group.pluginActions
-                    <IssueActionLinkButton
-                      size="zero"
-                      icon={integration.displayIcon}
-                      disabled={integration.disabled}
-                      tooltipProps={{
-                        title: integration.disabled
-                          ? integration.disabledText
-                          : undefined,
-                      }}
-                      onClick={() => {
-                        action.onClick();
-                        trackAnalytics('feedback.details-integration-issue-clicked', {
-                          organization,
-                          integration_key: integration.key,
-                        });
-                      }}
-                      href={action.href}
-                      external
-                    >
-                      <IssueActionName>{integration.displayName}</IssueActionName>
-                    </IssueActionLinkButton>
-                  ) : (
-                    <IssueActionButton
-                      {...sharedButtonProps}
-                      disabled={integration.disabled}
-                      tooltipProps={{
-                        title: integration.disabled
-                          ? integration.disabledText
-                          : undefined,
-                      }}
-                      onClick={() => {
-                        action.onClick();
-                        trackAnalytics('feedback.details-integration-issue-clicked', {
-                          organization,
-                          integration_key: integration.key,
-                        });
-                      }}
-                    />
-                  )}
-                </ErrorBoundary>
-              );
-            }
-
-            return (
-              <ErrorBoundary key={integration.key} mini>
-                <DropdownMenu
-                  trigger={triggerProps => (
-                    <IssueActionDropdownMenu
-                      {...sharedButtonProps}
-                      {...triggerProps}
-                      showChevron={false}
-                    />
-                  )}
-                  items={integration.actions.map(action => ({
-                    key: action.id,
-                    ...getActionLabelAndTextValue({
-                      action,
-                      integrationDisplayName: integration.displayName,
-                    }),
-                    onAction: action.onClick,
-                    disabled: integration.disabled,
-                  }))}
-                />
-              </ErrorBoundary>
-            );
-          })}
-        </IssueActionWrapper>
-      )}
-    </Fragment>
-  );
-}
-
-const IssueActionWrapper = styled('span')`
-  display: flex;
-  flex-wrap: wrap;
-  gap: ${p => p.theme.space.md};
-  line-height: 1.2;
-`;
-
-const LinkedIssue = styled(LinkButton)`
-  display: flex;
-  align-items: center;
-  padding: ${p => p.theme.space.xs} ${p => p.theme.space.sm};
-  border: none;
-  border-radius: ${p => p.theme.radius.md};
-  font-weight: normal;
-`;
-
-const IssueActionButton = styled(Button)`
-  display: flex;
-  align-items: center;
-  padding: ${p => p.theme.space.xs} ${p => p.theme.space.sm};
-  border: 1px dashed ${p => p.theme.tokens.border.primary};
-  border-radius: ${p => p.theme.radius.md};
-  font-weight: normal;
-`;
-
-const IssueActionLinkButton = styled(LinkButton)`
-  display: flex;
-  align-items: center;
-  padding: ${p => p.theme.space.xs} ${p => p.theme.space.sm};
-  border: 1px dashed ${p => p.theme.tokens.border.primary};
-  border-radius: ${p => p.theme.radius.md};
-  font-weight: normal;
-`;
-
-const IssueActionDropdownMenu = styled(DropdownButton)`
-  display: flex;
-  align-items: center;
-  padding: ${p => p.theme.space.xs} ${p => p.theme.space.sm};
-  border: 1px dashed ${p => p.theme.tokens.border.primary};
-  border-radius: ${p => p.theme.radius.md};
-  font-weight: normal;
-
-  &[aria-expanded='true'] {
-    border: 1px solid ${p => p.theme.tokens.border.primary};
-  }
-`;
-
-const IssueActionName = styled('div')`
-  display: block;
-  width: 100%;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  max-width: 200px;
-`;
-
-const LinkedIssueTooltipWrapper = styled('div')`
-  display: flex;
-  align-items: center;
-  gap: ${p => p.theme.space.xs};
-  white-space: nowrap;
-`;
-
-const LinkedIssueName = styled('div')`
-  display: block;
-  width: 100%;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  margin-right: ${p => p.theme.space['2xs']};
-`;
-
-const HorizontalSeparator = styled('div')`
-  width: 1px;
-  height: 14px;
-  /* eslint-disable-next-line @sentry/scraps/use-semantic-token */
-  background: ${p => p.theme.tokens.border.primary};
-`;
-
-const UnlinkButton = styled(Button)`
-  color: ${p => p.theme.tokens.content.secondary};
-`;
