+++ static/app/views/issueDetails/groupSidebar.tsx
@@ -8,7 +8,6 @@ import ErrorBoundary from 'sentry/components/errorBoundary';
 import {EventThroughput} from 'sentry/components/events/eventStatisticalDetector/eventThroughput';
 import {AssignedTo} from 'sentry/components/group/assignedTo';
 import type {OnAssignCallback} from 'sentry/components/group/assigneeSelector';
-import {ExternalIssueList} from 'sentry/components/group/externalIssuesList';
 import GroupReleaseStats from 'sentry/components/group/releaseStats';
 import {
   BACKEND_TAGS,
@@ -269,11 +268,6 @@ export function GroupSidebar({event, group, project, organization, environments}
           group={group}
           currentRelease={currentRelease}
         />
-      )}
-      {event && (
-        <ErrorBoundary mini>
-          <ExternalIssueList project={project} group={group} event={event} />
-        </ErrorBoundary>
       )}
       {renderPluginIssue()}
       {issueTypeConfig.pages.tagsTab.enabled && (
